\name{JavaVMRef-class}
\alias{JavaVMRef-class}
\title{JavaVMRef}
\description{Represents a pointer to the Java Virtual Machine}
\details{
	This Class is used to represent a pointer to the Java Vitrual Machine
}
\author{Steffen Durinck, \url{http://www.esat.kuleuven.ac.be/~sdurinck}} 

\keyword{methods}

