\name{MEXP14}
\docType{data}
\alias{MEXP14}
\alias{160miame2.txt.xml}
\alias{am2730miame}
\alias{am2731miame}
\alias{am2732m}
\alias{am2736m}
\alias{am2737m}
\alias{Flounder_Environmental_Tyne_vs_Alde_Male.xml}
\alias{tm1826m}
\alias{tm1827m}
\alias{tm1829m}
\alias{tm1830m}
\alias{tm1831m}

\title{Example dataset in MAGEML-format}

\description{This dataset is available from ArrayExpress(http://www.ebi.ac.uk/arrayexpress/) as E-MEXP-14 and contains data from a study for comparison of environmentally sampled flounder from the Tyne and Alde estuaries.
 }
\details{See ArrayExpress}
\references{Timothy Williams, Karl Gensberg, Steven Minchin and James Chipman (2003). A DNA expression array to detect toxic stress response in European flounder (Platichthys flesus). Aquatic Toxicology, 65, 141-157}


\keyword{datasets}
\source{ArrayExpress, http://www.ebi.ac.uk/arrayexpress/}
